requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./pu compile-reqs
#
-e file:.
attrs==25.4.0
aws-cdk-asset-awscli-v1==2.2.263
aws-cdk-asset-node-proxy-agent-v6==2.1.1
aws-cdk-cloud-assembly-schema==52.2.0
aws-cdk-lib==2.243.0
awscrt==0.31.2
black==26.3.1
boto3==1.42.69
boto3-stubs[dynamodb,lambda]==1.42.69
botocore[crt]==1.42.69
botocore-stubs==1.42.41
build==1.4.0
cachetools==7.0.5
cattrs==25.3.0
click==8.3.1
colorama==0.4.6
constructs==10.5.1
desert==2022.9.22
distlib==0.4.0
filelock==3.25.2
importlib-resources==6.5.2
isort==8.0.1
jmespath==1.1.0
jsii==1.127.0
librt==0.8.1
marshmallow==4.2.2
marshmallow-enum==1.5.1
marshmallow-union==0.1.15.post1
mypy==1.19.1
mypy-boto3-dynamodb==1.42.55
mypy-boto3-lambda==1.42.37
mypy-extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pip-tools==7.5.3
platformdirs==4.9.4
pluggy==1.6.0
publication==0.0.3
pyproject-api==1.10.0
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.1.3
pytokens==0.4.1
s3transfer==0.16.0
six==1.17.0
tomli-w==1.2.0
tox==4.49.1
typeguard==2.13.3
types-awscrt==0.31.3
types-s3transfer==0.16.0
typing-extensions==4.15.0
typing-inspect==0.9.0
urllib3==2.6.3
virtualenv==21.2.0
wheel==0.46.3

# The following packages are considered to be unsafe in a requirements file:
pip==26.0.1
setuptools==82.0.1
